mkpkg.sh update

+++ mkpkg.sh
@@ -1,10 +1,15 @@
 #!/bin/sh
 # Create makesbld package
 
+# if VERSION file not there, then quit
+
+[ ! -r VERSION ] && exit 1
+
 PKGNAME=makesbld
 VERSION=`cat VERSION`
 PKGDIR=/tmp
 BUILDDIR=${PKGDIR}/${PKGNAME}
+REV=1
 
 # copy files over to temp
 [ -d $BUILDDIR ] && rm -fr $BUILDDIR
@@ -20,9 +25,16 @@ do
 	sed -i -e "s^%%ver%%^$VERSION^" $file
 done 
 
+# rename doc dir
+mv -v usr/doc/${PKGNAME} usr/doc/${PKGNAME}-${VERSION}
+
+# gzip everything in man and doc
+gzip -v9 usr/doc/*/*
+gzip -v9 usr/man/*/*
+
 # make package
 
-/sbin/makepkg -c n -l n ${PKGDIR}/${PKGNAME}-${VERSION}-noarch-1.txz 
+/sbin/makepkg -c n -l n ${PKGDIR}/${PKGNAME}-${VERSION}-noarch-${REV}.txz 
 
 echo "Package $PKGNAME-$VERSION has been stored in $PKGDIR."
